requirements-dev.txt
# Testing
pytest>=7.4.0
pytest-cov>=4.1.0
pytest-mock>=3.11.0
pytest-asyncio>=0.21.0

# Code quality
black>=23.7.0
isort>=5.12.0
flake8>=6.1.0
mypy>=1.5.0
pylint>=2.17.0

# Type stubs
pandas-stubs>=2.0.0
types-requests>=2.31.0

# Documentation
sphinx>=7.0.0
sphinx-rtd-theme>=1.3.0

# Security
bandit>=1.7.5
safety>=2.3.5